+++ mini_projects/day-06-tuples.py
@@ -1,0 +1,12 @@
+days = (
+    "Monday",
+    "Tuesday",
+    "Wednesday",
+    "Thursday",
+    "Friday",
+    "Saturday",
+    "Sunday"
+)
+
+number = int(input("Enter the number of day :"))
+print(days[number-1])